// containers/main/src/components/app-menu/app-menu.types.ts
export interface IPKMenuItems {
  counter?: number;
  icon: string;
  activeIcon?: string;
  label: string;
  description?: string;
  url?: string;
  qaLabel?: string;
  roles: string[];
}

export interface IPKAppItems {
  icon: string;
  activeIcon?: string;
  label: string;
  description?: string;
  url: string;
  qaLabel?: string;
  roles?: string[];
}